+++ src/common/utils.ts
@@ -105,8 +105,7 @@ export const getCharacterLength = (str: string, maxCharacter?: number): { length
     };
   }
   let len = 0;
-  // eslint-disable-next-line no-plusplus
-  for (let i = 0; i < str.length; i++) {
+  for (let i = 0; i < str.length; i += 1) {
     let currentStringLength = 0;
     if (str.charCodeAt(i) > 127 || str.charCodeAt(i) === 94) {
       currentStringLength = 2;
@@ -133,7 +132,6 @@ export const getCharacterLength = (str: string, maxCharacter?: number): { length
   };
 };
 
-
 export const chunk = (arr: any[], size: number) =>
   Array.from({ length: Math.ceil(arr.length / size) }, (v, i) => arr.slice(i * size, i * size + size));
 
@@ -151,4 +149,3 @@ export const clone = (val) => {
   }
   return val;
 };
-